Rename RegularUserRegistrationRequestVo -> RegularUserRegistrationRequestDto
